% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_helpers.R
\name{ci_2_sd}
\alias{ci_2_sd}
\title{Convert Confidence Interval to Standard Deviation}
\usage{
ci_2_sd(upper_ci, lower_ci, n, interval = 95)
}
\arguments{
\item{upper_ci}{Upper confidence interval bound}

\item{lower_ci}{Lower confidence interval bound}

\item{n}{Sample size}

\item{interval}{Confidence interval level (90 or 95). Default is 95.}
}
\value{
Standard deviation
}
\description{
Converts confidence intervals to standard deviation using Cochrane methodology.
}
\examples{
# Convert 95\% CI to SD
ci_2_sd(upper_ci = 5.2, lower_ci = 2.8, n = 50)

# Convert 90\% CI to SD
ci_2_sd(upper_ci = 5.2, lower_ci = 2.8, n = 50, interval = 90)
}
\references{
Cochrane Handbook 5.1, Section 7.7.3.2
\url{https://handbook-5-1.cochrane.org/chapter_7/7_7_3_2_obtaining_standard_deviations_from_standard_errors_and.htm}
}
\seealso{
Other effect size functions: 
\code{\link{d_calc}()},
\code{\link{did_calculator}()},
\code{\link{sd_pooled}()},
\code{\link{se_2_sd}()},
\code{\link{var_d_calc}()}
}
\concept{effect size functions}
