% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_calc.R
\name{d_calc}
\alias{d_calc}
\title{Calculate Cohen's D or Glass's Delta}
\usage{
d_calc(stat_type, stat, sample_sd, n_t, n_c)
}
\arguments{
\item{stat_type}{Category of statistical result reported or derived from the paper.
Possible values are "d", "d_i_d", "d_i_m", "d_i_p", "eta_squared", "f_test",
"log_odds_ratio", "odds_ratio", "reg_coef", "s_m_d", "t_test", "unspecified null", "z".}

\item{stat}{Unstandardized effect size.}

\item{sample_sd}{Standard deviation of the relevant sample (preferably of the control group).}

\item{n_t}{Treatment group sample size.}

\item{n_c}{Control group sample size.}
}
\value{
Cohen's D or Glass's Delta value.
}
\description{
This function converts raw effect sizes into Cohen's D or Glass's Delta
based on the measure of sample variance for standardization.
}
\details{
All calculations derived from Cooper, Hedges, and Valentine (2009), except
for difference in proportions, which, to the best of our knowledge, Don Green
came up with while we were working on \emph{Prejudice Reduction: Progress and Challenges}.
We elaborate more on this estimator in the paper's appendix.
}
\examples{
# Example: Calculate d for a study that provides difference in means results
difference_in_means_result <- d_calc(stat_type = "d_i_m", stat = 1, sample_sd = 0.3)
# Example: Calculate d for a study that provides an F-test
f_test_result <- d_calc(stat_type = "f_test", stat = 1, n_t = 50, n_c = 40)
# Example: Use mapply to calculate d from rows in dataset
\dontrun{
library(dplyr)
BLPlabtools::contact_data |>
select(-d) |> # remove d in order to recalculate it
mutate(d = mapply(
FUN = d_calc,
stat_type = statistic,
stat =  unstand,
sample_sd = sd_c,
n_t = n_t,
n_c = n_c)) |> select(name_short, d)
}
}
\seealso{
\code{\link{var_d_calc}} for calculating the variance of effect sizes

Other effect size functions: 
\code{\link{ci_2_sd}()},
\code{\link{did_calculator}()},
\code{\link{sd_pooled}()},
\code{\link{se_2_sd}()},
\code{\link{var_d_calc}()}
}
\concept{effect size functions}
