% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_helpers.R
\name{did_calculator}
\alias{did_calculator}
\title{Calculate Difference-in-Differences Effect Size}
\usage{
did_calculator(
  mean_treatment_post,
  mean_treatment_pre,
  mean_control_post,
  mean_control_pre,
  sd
)
}
\arguments{
\item{mean_treatment_post}{Post-treatment mean for treatment group}

\item{mean_treatment_pre}{Pre-treatment mean for treatment group}

\item{mean_control_post}{Post-treatment mean for control group}

\item{mean_control_pre}{Pre-treatment mean for control group}

\item{sd}{Standard deviation (preferably from control group)}
}
\value{
Standardized difference-in-differences effect size
}
\description{
Computes a standardized difference-in-differences effect size from raw means.
}
\examples{
did_calculator(mean_treatment_post = 5.2, mean_treatment_pre = 4.8,
               mean_control_post = 4.5, mean_control_pre = 4.6, sd = 1.2)
}
\seealso{
Other effect size functions: 
\code{\link{ci_2_sd}()},
\code{\link{d_calc}()},
\code{\link{sd_pooled}()},
\code{\link{se_2_sd}()},
\code{\link{var_d_calc}()}
}
\concept{effect size functions}
