% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_robust.R
\name{map_robust}
\alias{map_robust}
\title{Perform robust meta-analysis (including in a pipe or split paradigm)}
\usage{
map_robust(x)
}
\arguments{
\item{x}{The dataset or subset to perform meta-analysis on.}
}
\value{
A tibble with meta-analysis results.
}
\description{
Performs robust meta-analysis on either data subsets or the whole dataset.
Its intended usage is in in context of the  split/map & pipes paradigm:
\code{dat |> split(~some_var) |> map(map_robust)}.
}
\details{
\code{dat |> map_robust()} will return the meta-analytic estimate for the whole dataset.
}
\note{
If there is only one study in a cluster, this function will
just reproduce that study's meta-analytic estimate verbatim. This is good for
some circumstances but not others. Adapt to your needs.
}
\examples{
# example 1: meta-analyze entire dataset
PaluckMetaSOP::sv_data |> map_robust()
# example 2: meta-analyze many subsets and create overall table
\dontrun{
library(dplyr); library(purrr)
sv_data |> split(~behavior_type) |> map(map_robust) |>
 bind_rows(.id = "behavior_type")
 }
}
\seealso{
\code{\link{d_calc}} and \code{\link{var_d_calc}} for preparing data before meta-analysis
}
