% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se.R
\name{robust_se}
\alias{robust_se}
\title{Robust Standard Error}
\usage{
robust_se(model, cluster)
}
\arguments{
\item{model}{A fitted linear model object (from lm())}

\item{cluster}{A vector indicating cluster membership for computing robust SEs}
}
\value{
A list with two elements: the robust covariance matrix and coefficient test results
}
\description{
Calculate cluster-robust standard errors for linear models
}
\author{
John-Henry Pezzuto
}
