% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_helpers.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Calculate Pooled Standard Deviation}
\usage{
sd_pooled(sd, n)
}
\arguments{
\item{sd}{Vector of standard deviations}

\item{n}{Vector of sample sizes (same length as sd)}
}
\value{
Pooled standard deviation
}
\description{
Pools standard deviations across multiple groups using Hedges (1981) method.
}
\examples{
sd_pooled(sd = c(2.1, 2.5, 1.9), n = c(30, 25, 35))
}
\references{
Hedges, L. V. (1981). Distribution Theory for Glass's Estimator of
Effect Size and Related Estimators. Journal of Educational Statistics, 6(2), 107-128.
}
\seealso{
Other effect size functions: 
\code{\link{ci_2_sd}()},
\code{\link{d_calc}()},
\code{\link{did_calculator}()},
\code{\link{se_2_sd}()},
\code{\link{var_d_calc}()}
}
\concept{effect size functions}
