% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_helpers.R
\name{se_2_sd}
\alias{se_2_sd}
\title{Convert Standard Error to Standard Deviation}
\usage{
se_2_sd(se, n)
}
\arguments{
\item{se}{Standard error}

\item{n}{Sample size}
}
\value{
Standard deviation
}
\description{
Simple conversion from standard error to standard deviation.
}
\examples{
se_2_sd(se = 0.5, n = 100)
}
\seealso{
Other effect size functions: 
\code{\link{ci_2_sd}()},
\code{\link{d_calc}()},
\code{\link{did_calculator}()},
\code{\link{sd_pooled}()},
\code{\link{var_d_calc}()}
}
\concept{effect size functions}
