% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lm.R
\name{summarise_lm}
\alias{summarise_lm}
\title{Print Core Results from summary(lm()) in a Neat Table}
\usage{
summarise_lm(dataset, y = NULL, x = NULL, coefs_only = TRUE, dgts = 5)
}
\arguments{
\item{dataset}{The dataset to be used for regression analysis.}

\item{y}{The response variable in the linear model.}

\item{x}{The predictor variable in the linear model.}

\item{coefs_only}{Logical, indicating whether to display only the coefficients table (default: TRUE).}

\item{dgts}{Number of digits for rounding coefficients (default: 5).}
}
\value{
If coefs_only is TRUE, returns a rounded coefficients table; otherwise, returns the full summary.
}
\description{
\code{summarise_lm} takes the built in R functions \code{summary(lm())}
and lets you use them in a sequence of pipes.
In particular this is useful when you want to test for
the magnitude of a relationship between two variables in
many different subsets of data at once.
}
\note{
The function defaults to using \code{d} as the response variable and \code{se_d} as the predictor
if these columns exist in the dataset. You can override these by specifying \code{y} and \code{x} parameters.
}
\examples{
# example 1: entire dataset
PaluckTools::contact_data |> summarise_lm()
# example 2: split and apply to many subsets
\dontrun{
library(purrr)
PaluckTools::sv_data |> split(~study_design) |> map(summarise_lm)
}
}
\seealso{
\code{\link{summarise_table}} for frequency tables, \code{\link{study_count}} for counting studies

Other summary functions: 
\code{\link{study_count}()},
\code{\link{summarise_table}()}
}
\concept{summary functions}
