% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_table.R
\name{summarise_table}
\alias{summarise_table}
\title{Create a Frequency Table for a Variable in a Data Frame}
\usage{
summarise_table(data, var_name)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{var_name}{The name of the variable/column to generate the frequency table.}
}
\value{
A table showing the frequency of each unique value in the specified variable.
}
\description{
\code{summarise_table} takes a built in function (table) and turns it into something
that can be integrated into a sequence of pipes. As with \code{summarise_lm},
this is useful for when you want to see many tables at once,
or put a few tables into a bigger table.
Writing this was surprisingly complicated -- it uses enquo() to capture
the unquoted variable name and !! to unquote it within the function.
}
\examples{
# simple example: entire dataset
PaluckTools::sv_data |> summarise_table(behavior_type)
\dontrun{
# example with split and apply to many subsets
library(purrr)
PaluckTools::sv_data |> split(~study_design) |>
map(~ summarise_table(., behavior_type)) |> bind_rows(.id = "study_design")
}
}
\seealso{
\code{\link{summarise_lm}} for regression summaries, \code{\link{study_count}} for counting studies

Other summary functions: 
\code{\link{study_count}()},
\code{\link{summarise_lm}()}
}
\concept{summary functions}
