% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dockerfile.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Convert the output from \code{sessionInfo()} into the contents of a Dockerfile}
\usage{
write_dockerfile(
  write_file = TRUE,
  org = "rocker",
  image = NULL,
  apt_packages = TRUE,
  dir = "./",
  filename = "Dockerfile",
  all_packages = FALSE
)
}
\arguments{
\item{write_file}{Defaults to TRUE, which will write a dockerfile. Else,
output will be written to console (or whatever connection is open).
console (or whatever connection you have open). The dockerfile's name
will combine the dir and filename parameters: by default, './Dockerfile'.
Writing a file will overwrite any Dockerfile in the same location.
'TRUE' is probably appropriate for exporting to a tool like mybinder.org;
'FALSE ' is appropriate if you already have a Dockerfile and you prefer
to paste sections from this function into it.}

\item{org}{Defaults to Rocker (https://www.rocker-project.org/); another
option is Docker's official R images (https://hub.docker.com/_/r-base).
For more on Rocker, see https://arxiv.org/abs/1710.03675.}

\item{image}{Defaults to automatically detect R version and use matching rocker/r-base image.
Available tags for r-base are listed here: https://hub.docker.com/r/rocker/r-base/tags.
Other Rocker images can be browsed here: https://hub.docker.com/u/rocker.}

\item{apt_packages}{should standard needed apt packages be installed?
Defaults to true and installs libcurl4-openssl-dev, libssl-dev, and
libxml2-dev.}

\item{dir}{Defaults to current directory, './'.}

\item{filename}{Defaults to 'Dockerfile', making the composite file
'./Dockerfile'.}

\item{all_packages}{If TRUE, includes all installed packages. If FALSE (default),
only includes currently loaded packages.}
}
\description{
Convert the output from \code{sessionInfo()} into the contents of a Dockerfile
}
\examples{
write_dockerfile()
write_dockerfile(write_file = FALSE, image = 'r-base:4.4.1')
write_dockerfile(all_packages = TRUE) # Include all installed packages
}
